refactor: resolveGitDir uses execGit
