+++ src/gitWorkflow.js
@@ -1,32 +1,14 @@
 'use strict'
 
-const path = require('path')
-const execa = require('execa')
 const gStatus = require('g-status')
 const del = require('del')
 const debug = require('debug')('lint-staged:git')
 
+const execGit = require('./execGit')
+
 let workingCopyTree = null
 let indexTree = null
 let formattedIndexTree = null
-
-function getAbsolutePath(dir) {
-  return path.isAbsolute(dir) ? dir : path.resolve(dir)
-}
-
-async function execGit(cmd, options) {
-  const { cwd } = options
-  debug('Running git command', cmd)
-  try {
-    const { stdout } = await execa('git', [].concat(cmd), {
-      ...options,
-      cwd: getAbsolutePath(cwd)
-    })
-    return stdout
-  } catch (err) {
-    throw new Error(err)
-  }
-}
 
 async function writeTree(options) {
   return execGit(['write-tree'], options)
